fixed find by id issue

+++ routes/organisations.js
@@ -1,5 +1,6 @@
 const { Organisation, validate } = require('../models/organisation')
 const express = require('express')
+const mongoose = require('mongoose')
 const router = express.Router()
 
 router.get('/', async (req, res) => {
@@ -27,32 +28,37 @@ router.put('/:id', async (req, res) => {
   const { error } = validate(req.body)
   if (error) return res.status(400).send(error.details[0].message)
 
-  const organisation = await Organisation.findByIdAndUpdate(req.params.id, {
-    name: req.body.name,
-    email: req.body.email,
-    city: req.body.city,
-    sectors: req.body.sectors
-  }, { new: true })
+  if(mongoose.Types.ObjectId.isValid(req.params.id)) {
+    const organisation = await Organisation.findByIdAndUpdate(req.params.id, {
+      name: req.body.name,
+      email: req.body.email,
+      city: req.body.city,
+      sectors: req.body.sectors
+    }, { new: true })
 
-  if (!organisation) return res.status(404).send('The organisation with the given ID was not found.')
-
-  res.send(organisation)
+    res.send(organisation)
+  } else {
+    return res.status(404).send('The organisation with the given ID was not found.')
+  }
 })
 
 router.delete('/:id', async (req, res) => {
-  const organisation = await Organisation.findByIdAndRemove(req.params.id)
+  if(mongoose.Types.ObjectId.isValid(req.params.id)) {
+    const organisation = await Organisation.findByIdAndRemove(req.params.id)
+    res.send(organisation)
+  } else {
+    return res.status(404).send('The organisation with the given ID was not found.')
+  }
 
-  if (!organisation) return res.status(404).send('The organisation with the given ID was not found.')
-
-  res.send(organisation)
 })
 
 router.get('/:id', async (req, res) => {
-  const organisation = await Organisation.findById(req.params.id)
-
-  if (!organisation) return res.status(404).send('The organisation with the given ID was not found.')
-
-  res.send(organisation)
+  if(mongoose.Types.ObjectId.isValid(req.params.id)) {
+    const organisation = await Organisation.findById(req.params.id)
+    res.send(organisation)
+  } else {
+    return res.status(404).send('The organisation with the given ID was not found.')
+  }
 })
 
 module.exports = router
